Move app files back to src due to error
